LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/crossorigin-expected.txt

CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: ReferenceError: Can't find variable: nonExistentMethod
CONSOLE MESSAGE: ReferenceError: Can't find variable: nonExistentMethod
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
html-script-module-crossOrigin



FAIL Root module, Error in CORS-same-origin script assert_equals: Unexpected _log value expected "running,8-1" but got "running,8-18"
PASS Root module, Blocked script download, missing CORS ACAO header
PASS Root module, Blocked script download, crossorigin attribute with missing CORS ACAO header
PASS Root module, Blocked script download, mismatched CORS ACAO header
FAIL Imported module, Error in CORS-same-origin script assert_equals: Unexpected _log value expected "running,8-1" but got "running,8-18"
PASS Imported module, Blocked script download, missing CORS ACAO header
PASS Imported module, Blocked script download, crossorigin attribute with missing CORS ACAO header
PASS Imported module, Blocked script download, mismatched CORS ACAO header
